Fix compilation under mingw64; wvtestmain runs under wine64.
The bad news is there are lots of test failures.  The good news is
there are also quite a lot of test passes, so it's probably fairly
tractable.

+++ utils/t/wvtask.t.cc
@@ -10,7 +10,7 @@ long glob;
 
 static void gentask(void *userdata)
 {
-    long startval = (long)userdata, val = startval;
+    long startval = (long)(long long)userdata, val = startval;
     
     while (val - startval < 3)
     {
